chore: Rename utils to docutil
Closes #17
