+++ pkg/docutil/encoder.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package utils
+package docutil
 
 import "encoding/base64"
